SensorReading refactor and changes to GeneralSensorConfig:
The SensorReading object now points to a GeneralSensorConfig instead of itself recreating all the values as field.
This makes creating SensorReading objects much shorter.
Further changed handling of 'values' in SensorReading to be easier accessible and tied tighter to the parameter (names)

In GeneralSensorConfig implemented a map for the sensor Definition.